+++ src/Simplex/Chat/Types.hs
@@ -1629,7 +1629,7 @@ instance ToJSON FileType where
   toJSON = J.String . textEncode
   toEncoding = JE.text . textEncode
 
-data FileProhibited = FileProhibited {maxSize :: Int64, badgeStatus :: Maybe BadgeStatus}
+data FileProhibited = FileProhibited {maxSize :: Integer, badgeStatus :: Maybe BadgeStatus}
   deriving (Eq, Show)
 
 data RcvFileTransfer = RcvFileTransfer
